requirements.txt
snscrape
readability-lxml
requests
beautifulsoup4
python-dotenv
PyYAML
